refactor(wdl-lint)!: promote `Deprecated*` rules
