+++ crates/wdl-analysis/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* `DeprecatedObject`, `DeprecatedPlaceholder`, and `DeprecatedRuntimeSection` rules, which ensure that deprecated
+  language features are not used ([#1166](https://github.com/stjude-rust-labs/sprocket/pull/1166)).
+
 ## 0.25.0 - 2026-08-26
 
 #### Added
