+++ crates/wdl-analysis/src/config.rs
@@ -15,6 +15,9 @@ use wdl_ast::SupportedVersion;
 use wdl_ast::SyntaxNode;
 
 use crate::CommandSectionIndentationRule;
+use crate::DeprecatedObjectRule;
+use crate::DeprecatedPlaceholderRule;
+use crate::DeprecatedRuntimeSectionRule;
 use crate::ExceptDirectiveValidRule;
 use crate::Exceptable as _;
 use crate::FormatConfig;
@@ -353,6 +356,21 @@ pub struct DiagnosticsConfig {
     /// A value of `None` disables the diagnostic.
     #[toml(FromToml with = parse_string)]
     pub command_section_indentation: Option<Severity>,
+    /// The severity for the deprecated `object` diagnostic.
+    ///
+    /// A value of `None` disables the diagnostic.
+    #[toml(FromToml with = parse_string)]
+    pub deprecated_object: Option<Severity>,
+    /// The severity for the deprecated placeholder option diagnostic.
+    ///
+    /// A value of `None` disables the diagnostic.
+    #[toml(FromToml with = parse_string)]
+    pub deprecated_placeholder: Option<Severity>,
+    /// The severity for the deprecated `runtime` section diagnostic.
+    ///
+    /// A value of `None` disables the diagnostic.
+    #[toml(FromToml with = parse_string)]
+    pub deprecated_runtime_section: Option<Severity>,
 }
 
 impl Default for DiagnosticsConfig {
@@ -375,6 +393,9 @@ impl DiagnosticsConfig {
         let mut known_rules = None;
         let mut except_directive_valid = None;
         let mut command_section_indentation = None;
+        let mut deprecated_object = None;
+        let mut deprecated_placeholder = None;
+        let mut deprecated_runtime_section = None;
 
         for rule in rules {
             let rule = rule.as_ref();
@@ -393,6 +414,11 @@ impl DiagnosticsConfig {
                 ExceptDirectiveValidRule::ID => except_directive_valid = Some(rule.severity()),
                 CommandSectionIndentationRule::ID => {
                     command_section_indentation = Some(rule.severity())
+                }
+                DeprecatedObjectRule::ID => deprecated_object = Some(rule.severity()),
+                DeprecatedPlaceholderRule::ID => deprecated_placeholder = Some(rule.severity()),
+                DeprecatedRuntimeSectionRule::ID => {
+                    deprecated_runtime_section = Some(rule.severity())
                 }
                 unrecognized => {
                     warn!(unrecognized, "unrecognized rule");
@@ -415,6 +441,9 @@ impl DiagnosticsConfig {
             known_rules,
             except_directive_valid,
             command_section_indentation,
+            deprecated_object,
+            deprecated_placeholder,
+            deprecated_runtime_section,
         }
     }
 
@@ -436,6 +465,9 @@ impl DiagnosticsConfig {
                 KnownRulesRule::ID => self.known_rules = None,
                 ExceptDirectiveValidRule::ID => self.except_directive_valid = None,
                 CommandSectionIndentationRule::ID => self.command_section_indentation = None,
+                DeprecatedObjectRule::ID => self.deprecated_object = None,
+                DeprecatedPlaceholderRule::ID => self.deprecated_placeholder = None,
+                DeprecatedRuntimeSectionRule::ID => self.deprecated_runtime_section = None,
                 _ => {}
             }
         }
@@ -457,6 +489,9 @@ impl DiagnosticsConfig {
             known_rules: None,
             except_directive_valid: None,
             command_section_indentation: None,
+            deprecated_object: None,
+            deprecated_placeholder: None,
+            deprecated_runtime_section: None,
         }
     }
 }
