+++ crates/wdl-analysis/tests/analysis/cannot-coerce-string/source.diagnostics
@@ -1,5 +1,29 @@
+warning[DeprecatedPlaceholder]: use of the deprecated `${}` placeholder option
+   ┌─ tests/analysis/cannot-coerce-string/source.wdl:10:21
+   │
+10 │     String x = "foo ${a}"
+   │                     ^
+   │
+   = fix: replace the opening token `$` with `~`
+
 error: cannot coerce type `Array[String]` to type `String`
    ┌─ tests/analysis/cannot-coerce-string/source.wdl:10:23
    │
 10 │     String x = "foo ${a}"
    │                       ^ this is type `Array[String]`
+
+warning[DeprecatedPlaceholder]: use of the deprecated `${}` placeholder option
+   ┌─ tests/analysis/cannot-coerce-string/source.wdl:13:21
+   │
+13 │     String y = "foo ${sep="," a}"
+   │                     ^
+   │
+   = fix: replace the opening token `$` with `~`
+
+warning[DeprecatedPlaceholder]: use of the deprecated `sep` placeholder option
+   ┌─ tests/analysis/cannot-coerce-string/source.wdl:13:23
+   │
+13 │     String y = "foo ${sep="," a}"
+   │                       ^^^^^^^
+   │
+   = fix: replace the `sep` placeholder option with a call to the `sep()` standard library function
