+++ crates/wdl-analysis/tests/analysis/deprecated-placeholder-options-v1.0/source.diagnostics
@@ -1,0 +1,31 @@
+warning[UnusedDeclaration]: unused declaration `bad_sep_option`
+   ┌─ tests/analysis/deprecated-placeholder-options-v1.0/source.wdl:14:12
+   │
+14 │     String bad_sep_option = "~{sep="," numbers}"
+   │            ^^^^^^^^^^^^^^
+
+warning[UnusedDeclaration]: unused declaration `bad_true_false_option`
+   ┌─ tests/analysis/deprecated-placeholder-options-v1.0/source.wdl:15:12
+   │
+15 │     String bad_true_false_option = "~{true="--enable-foo" false="" allow_foo}"
+   │            ^^^^^^^^^^^^^^^^^^^^^
+
+warning[UnusedDeclaration]: unused declaration `bad_default_option`
+   ┌─ tests/analysis/deprecated-placeholder-options-v1.0/source.wdl:16:12
+   │
+16 │     String bad_default_option = "~{default="false" bar}"
+   │            ^^^^^^^^^^^^^^^^^^
+
+warning[UnusedDeclaration]: unused declaration `bad_interpolation_option`
+   ┌─ tests/analysis/deprecated-placeholder-options-v1.0/source.wdl:17:12
+   │
+17 │     String bad_interpolation_option = "${bar}"
+   │            ^^^^^^^^^^^^^^^^^^^^^^^^
+
+warning[DeprecatedPlaceholder]: use of the deprecated `${}` placeholder option
+   ┌─ tests/analysis/deprecated-placeholder-options-v1.0/source.wdl:17:40
+   │
+17 │     String bad_interpolation_option = "${bar}"
+   │                                        ^
+   │
+   = fix: replace the opening token `$` with `~`
