+++ crates/wdl-analysis/tests/analysis/deprecated-runtime-section/source.diagnostics
@@ -1,0 +1,7 @@
+warning[DeprecatedRuntimeSection]: task `deprecated_runtime` contains a deprecated `runtime` section
+  ┌─ tests/analysis/deprecated-runtime-section/source.wdl:6:5
+  │
+6 │     runtime {}
+  │     ^^^^^^^
+  │
+  = fix: replace the `runtime` section with a `requirements` section
