+++ crates/wdl-analysis/tests/analysis/deprecated-runtime-section/source.wdl
@@ -1,5 +1,3 @@
-#@ except: RequirementsSection, MetaDescription, EmptyOutputs, BashSetSyntax, MetaSections
-
 version 1.3
 
 task deprecated_runtime {
