+++ crates/wdl-analysis/tests/analysis/placeholder-options/source.wdl
@@ -1,6 +1,6 @@
 ## Test cases for placeholder options analysis
 
-#@ except: UnusedDeclaration
+#@ except: UnusedDeclaration, DeprecatedPlaceholder
 
 version 1.3
