+++ crates/wdl-lint/RULES.md
@@ -12,9 +12,6 @@ This table documents all `sprocket` lint rules implemented on the `main` branch 
 | `ContainerUri`              | Clarity, Portability                                        | Ensures that values for the `container` key within `runtime`/`requirements` sections are well-formed.                                       |                                                         |
 | `DeclarationName`           | Naming, Style, Clarity                                      | Ensures declaration names do not redundantly include their type name.                                                                       |                                                         |
 | `DenyGlobStar`              | Correctness, Clarity                                        | Ensures outputs do not contain `glob("*")`.                                                                                                 |                                                         |
-| `DeprecatedObject`          | Deprecated                                                  | Ensures that the deprecated `Object` types are not used.                                                                                    |                                                         |
-| `DeprecatedPlaceholder`     | Deprecated                                                  | Ensures that deprecated expression placeholder options are not used.                                                                        |                                                         |
-| `DeprecatedRuntimeSection`  | Deprecated                                                  | Ensures that deprecated `runtime` sections are not used in WDL v1.2 and later.                                                                        |                                                         |
 | `DescriptionLength`         | SprocketCompatibility                                       | Ensures that description meta entries are not too long for display in Sprocket documentation                                                |                                                         |
 | `DocCommentTabs`            | Style, Clarity                                              | Ensures that doc comments do not contain tab characters.                                                                                    |                                                         |
 | `DocMetaStrings`            | SprocketCompatibility                                       | Ensures that reserved meta keys have string values for wdl-doc compatibility                                                                |                                                         |
