+++ crates/wdl-lint/src/lib.rs
@@ -141,9 +141,7 @@ pub fn rules(config: &Config) -> Vec<Box<dyn Rule + Send + Sync>> {
         Box::<rules::MetaSectionsRule>::default(),
         Box::<rules::CallInputKeywordRule>::default(),
         Box::<rules::SectionOrderingRule>::default(),
-        Box::<rules::DeprecatedObjectRule>::default(),
         Box::<rules::MetaDescriptionRule>::default(),
-        Box::<rules::DeprecatedPlaceholderRule>::default(),
         Box::new(rules::ExpectedRuntimeKeysRule::new(config)),
         Box::<rules::EmptyDocCommentRule>::default(),
         Box::<rules::DocMetaStringsRule>::default(),
@@ -156,7 +154,6 @@ pub fn rules(config: &Config) -> Vec<Box<dyn Rule + Send + Sync>> {
         Box::<rules::HostPathLiteralsRule>::default(),
         Box::<rules::ContainerUriRule>::default(),
         Box::<rules::RequirementsSectionRule>::default(),
-        Box::<rules::DeprecatedRuntimeSectionRule>::default(),
         Box::<rules::ParameterDescriptionRule>::default(),
         Box::<rules::ConciseInputRule>::default(),
         Box::<rules::ShellCheckRule>::default(),
