+++ crates/wdl-lint/src/rules.rs
@@ -6,9 +6,6 @@ mod concise_input;
 mod container_uri;
 mod declaration_name;
 mod deny_glob_star;
-mod deprecated_object;
-mod deprecated_placeholder;
-mod deprecated_runtime_section;
 mod description_length;
 mod doc_comment_tabs;
 mod doc_meta_strings;
@@ -43,9 +40,6 @@ pub use concise_input::*;
 pub use container_uri::*;
 pub use declaration_name::*;
 pub use deny_glob_star::*;
-pub use deprecated_object::*;
-pub use deprecated_placeholder::*;
-pub use deprecated_runtime_section::*;
 pub use description_length::*;
 pub use doc_comment_tabs::*;
 pub use doc_meta_strings::*;
