--- crates/wdl-lint/tests/lints/deprecated-object/source.errors
@@ -1,24 +1,0 @@
-note[DeprecatedObject]: use of a deprecated `Object` type
-   ┌─ tests/lints/deprecated-object/source.wdl:11:9
-   │
-11 │         Object an_unbound_literal_object
-   │         ^^^^^^
-   │
-   = fix: replace the `Object` with a `Map` or a `Struct`
-
-note[DeprecatedObject]: use of a deprecated `Object` type
-   ┌─ tests/lints/deprecated-object/source.wdl:14:5
-   │
-14 │     Object a_bound_literal_object = object {
-   │     ^^^^^^
-   │
-   = fix: replace the `Object` with a `Map` or a `Struct`
-
-note[DeprecatedObject]: use of a deprecated `Object` type
-   ┌─ tests/lints/deprecated-object/source.wdl:20:9
-   │
-20 │         Object another_bound_literal_object = object {
-   │         ^^^^^^
-   │
-   = fix: replace the `Object` with a `Map` or a `Struct`
-
