--- crates/wdl-lint/tests/lints/deprecated-placeholder-options-v1.0/source.errors
@@ -1,8 +1,0 @@
-note[DeprecatedPlaceholder]: use of the deprecated `${}` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.0/source.wdl:20:40
-   │
-20 │     String bad_interpolation_option = "${bar}"
-   │                                        ^
-   │
-   = fix: replace the opening token `$` with `~`
-
