--- crates/wdl-lint/tests/lints/deprecated-placeholder-options-v1.1/source.errors
@@ -1,72 +1,0 @@
-note[DeprecatedPlaceholder]: use of the deprecated `sep` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:15:32
-   │
-15 │     String bad_sep_option = "~{sep="," numbers}"
-   │                                ^^^^^^^
-   │
-   = fix: replace the `sep` placeholder option with a call to the `sep()` standard library function
-
-note[DeprecatedPlaceholder]: use of the deprecated `true`/`false` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:16:39
-   │
-16 │     String bad_true_false_option = "~{true="--enable-foo" false="" allow_foo}"
-   │                                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-   │
-   = fix: replace the `true`/`false` placeholder option with an `if`/`else` expression
-
-note[DeprecatedPlaceholder]: use of the deprecated `default` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:17:36
-   │
-17 │     String bad_default_option = "~{default="false" bar}"
-   │                                    ^^^^^^^^^^^^^^^
-   │
-   = fix: replace the `default` placeholder option with a call to the `select_first()` standard library function
-
-note[DeprecatedPlaceholder]: use of the deprecated `${}` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:18:40
-   │
-18 │     String bad_interpolation_option = "${bar}"
-   │                                        ^
-   │
-   = fix: replace the opening token `$` with `~`
-
-note[DeprecatedPlaceholder]: use of the deprecated `${}` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:19:31
-   │
-19 │     String bad_dual_option = "${sep="," numbers}"
-   │                               ^
-   │
-   = fix: replace the opening token `$` with `~`
-
-note[DeprecatedPlaceholder]: use of the deprecated `sep` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:19:33
-   │
-19 │     String bad_dual_option = "${sep="," numbers}"
-   │                                 ^^^^^^^
-   │
-   = fix: replace the `sep` placeholder option with a call to the `sep()` standard library function
-
-note[DeprecatedPlaceholder]: use of the deprecated `sep` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:22:28
-   │
-22 │         python script.py ~{sep=" " numbers}
-   │                            ^^^^^^^
-   │
-   = fix: replace the `sep` placeholder option with a call to the `sep()` standard library function
-
-note[DeprecatedPlaceholder]: use of the deprecated `true`/`false` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:23:27
-   │
-23 │         example-command ~{true="--enable-foo" false="" allow_foo}
-   │                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-   │
-   = fix: replace the `true`/`false` placeholder option with an `if`/`else` expression
-
-note[DeprecatedPlaceholder]: use of the deprecated `default` placeholder option
-   ┌─ tests/lints/deprecated-placeholder-options-v1.1/source.wdl:24:27
-   │
-24 │         another-command ~{default="foobar" bar}
-   │                           ^^^^^^^^^^^^^^^^
-   │
-   = fix: replace the `default` placeholder option with a call to the `select_first()` standard library function
-
