--- crates/wdl-lint/tests/lints/runtime-keys-wdl-1.2/source.errors
@@ -1,8 +1,0 @@
-note[DeprecatedRuntimeSection]: task `a_task_with_no_keys` contains a deprecated `runtime` section
-   ┌─ tests/lints/runtime-keys-wdl-1.2/source.wdl:13:5
-   │
-13 │     runtime {}
-   │     ^^^^^^^
-   │
-   = fix: replace the `runtime` section with a `requirements` section
-
